feat(api)!: checked argument accessors, and argc as usize
Every native opened by hand-checking its arguments: an `argc != N` test
the VM had already performed before dispatching, a type predicate, then
an unchecked cast. Roughly forty copies of that, and each failure
collapsed to the same generic message.

oak_arg_i32/f32/number/bool/cstring/string/native/self read one argument
and either yield it or raise an error naming the binding, the parameter
position, and the expected versus actual type. The whole stdlib and the
compiler builtins now go through them, dropping the dead arity checks and
the three separate copies of number_as_f32.

oak_arg_native is the one with teeth: oak_native_instance only asserts
"some native record" before reinterpreting whatever C struct is behind
it, so a receiver of the wrong bound type was silently misread. That is
reachable from C through oak_vm_call, where no compile-time check
applies. oak_arg_self applies it to the receiver of an instance method.

argc is now usize throughout the native ABI -- it is a count, and the
accessors index with it. oak_vm_call and oak_native_args_match follow.
The builtin-method prototypes move from the top of the table that uses
them into the internal header, where they cannot drift from the callback
contract; they already had.

Tests that asserted "native function 'x' failed" now assert the actual
reason, which is the point.

Breaking: oak_native_fn_t, oak_attr_runtime_cb_t, oak_vm_call and
oak_native_args_match all take usize counts.

+++ include/oak_value.h
@@ -527,7 +527,7 @@ typedef oak_fn_call_result_t (*oak_attr_runtime_cb_t)(
     oak_native_call_t* call,
     const char* fn_name,
     const oak_value_t* args,
-    int argc,
+    usize argc,
     void* user_data);
 
 struct oak_attr_hook_entry
@@ -536,10 +536,19 @@ struct oak_attr_hook_entry
   void* ud;
 };
 
+/* A bound C function.  `args` is a borrowed view into the VM stack: the callee
+ * must not decref an argument, and must incref anything it keeps past the call.
+ * The value written to *out_result is *moved* to the VM -- hand back a fresh
+ * reference (which every oak_vm_*_new constructor gives you) and never assign
+ * an object argument straight through, because the VM decrefs the arguments
+ * after taking the result.  Leaving *out_result untouched yields none.
+ *
+ * `argc` always equals the arity the binding was registered with; the VM
+ * rejects a mismatch before dispatching, so a callback need not re-check it. */
 typedef oak_fn_call_result_t (*oak_native_fn_t)(
     oak_native_call_t* call,
     const oak_value_t* args,
-    int argc,
+    usize argc,
     oak_value_t* out_result);
 
 typedef struct oak_obj_native_fn oak_obj_native_fn_t;
